// src/shared/api/models/postApiMobilepayCreateMobilepayOrderBody.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiMobilepayCreateMobilepayOrderBodyCustomerId } from './postApiMobilepayCreateMobilepayOrderBodyCustomerId';

export type PostApiMobilepayCreateMobilepayOrderBody = {
    /** @minLength 8 */
    phoneNumber: string;
    currency: string;
    /** */
    amount: number;
    reference: string;
    description: string;
    customerId: PostApiMobilepayCreateMobilepayOrderBodyCustomerId;
};